+++ CHANGES.md
@@ -4,7 +4,6 @@
 Changes in 12.9
 ---------------
 - Fix PHP warning undefined index when school was deleted in School.php & Schools.php
-- Security fix Request Forgery: remove modfunc= in links in Modules.php & MarkDownHTML.fnc.php
 - Performance: do not sort `functions/*.php` files in Warehouse.php
 - SQL performance: replace subqueries with LEFT JOINs in StaffBalances.php, StudentSummary.php, StudentBalances.php & Export.php
 - Use Select2 input when more than 12 Grade Levels in GetStuList.fnc.php & Search.fnc.php
@@ -19,6 +18,7 @@ Changes in 12.9
 - Use button() with "+" icon instead of text link in MassDrops.php, MassRequests.php & MassSchedule.php
 - Fix skip begin / end dates / number if "No Value" checked in CustomFields.fnc.php
 - Security fix #372 forbid path traversal sequences in FileUpload.fnc.php, thanks to @kazisabu
+- Security fix #371 Request Forgery add CSRF token to links/forms containing `modfunc=` in PreparePHP_SELF.fnc.php, Warehouse.php & index.php, thanks to @kazisabu
 
 Changes in 12.8.1
 -----------------